Fix the csv file used to test the results (probably created with an old proj4 version). UTM30 and 31 now pass if compared with proj4  v4.8.0. There  is still a problem with UTM32 but CTS is consistent with Circe (I changed the epsg file just to pass the test - not very clean). Next line to block is line 24. I refactored the test packages (op is for all kind of operation, op.projection for mathematical projections and op.transformation for operations including a datum change). Now, the csv resource file should match the class file using it.
