+++ src/test/java/org/cts/op/EPSGTransform.java
@@ -29,12 +29,13 @@
  *
  * For more information, please consult: <https://github.com/irstv/cts/>
  */
-package org.cts.op.transformation;
+package org.cts.op;
 
 import org.cts.IllegalCoordinateException;
 import org.cts.crs.CRSException;
 import org.cts.crs.CoordinateReferenceSystem;
 import org.cts.crs.GeodeticCRS;
+import org.cts.op.BaseCoordinateTransformTest;
 import org.cts.parser.prj.PrjWriter;
 import static org.junit.Assert.assertTrue;
 import org.junit.Test;
