+++ src/test/resources/org/cts/op/crstransform.csv
@@ -1,0 +1,38 @@
+id;csNameSrc;csNameSrc_X;csNameSrc_Y;csNameDest;csNameDest_X;csNameDest_Y;tolerance
+1;EPSG:4326;2.114551393;50.345609791;EPSG:2154;636890.74032145;7027895.26344997;0.01
+2;EPSG:4326;2.114551393;50.345609791;EPSG:27582;584173.736059813;2594514.82833411;0.01
+3;EPSG:4326;2.114551393;50.345609791;EPSG:27572;584173.736059813;2594514.82833411;0.01
+4;EPSG:4326;2.114551393;50.345609791;EPSG:27561;584202.918015345;294079.132285403;0.01
+5;EPSG:4326;2.114551393;50.345609791;EPSG:27561;584202.918015345;294079.132285403;0.01
+6;EPSG:4326;2.114551393;50.345609791;EPSG:27562;584173.736059813;594514.828334111;0.01
+7;EPSG:4326;2.114551393;50.345609791;EPSG:27563;584106.984970143;895712.712203558;0.01
+8;EPSG:4326;2.114551393;50.345609791;EPSG:27564;-15731.7949147033;1098366.1177769;0.01
+9;EPSG:4326;2.114551393;50.345609791;EPSG:27571;584202.918015345;1294079.1322854;0.01
+10;EPSG:4326;2.114551393;50.345609791;EPSG:27573;584106.984970143;3895712.71220356;0.01
+11;EPSG:4326;2.114551393;50.345609791;EPSG:27574;-15731.7949147033;5098366.1177769;0.01
+# D'où vient ce calcul ? Avec proj4.8.0, ci-dessous, ca marche
+#12;EPSG:4326;2.114551393;50.345609791;EPSG:23030;863838.2127826;5589689.13239247;0.01
+12;EPSG:4326;2.114551393;50.345609791;EPSG:23030;863926.128694867;5589792.87595142;0.01
+# D'où vient ce calcul ? Avec proj4.8.0, sur la ligne d'après, ca marche
+#13;EPSG:4326;2.114551393;50.345609791;EPSG:23031;436996.012761202;5577546.16879576;0.01
+13;EPSG:4326;2.114551393;50.345609791;EPSG:23031;437091.854358749;5577642.33578973;0.01
+# D'où vient ce calcul ?
+#14;EPSG:4326;2.114551393;50.345609791;EPSG:23032;10283.3249679306;5599876.08211079;0.01
+# Celui-ci ne marche pas non plus avec proj4.8.0
+# J'ai l'impression qu'il y a confusion sur la zone UTM (qui de toutes façons n'est pas la bonne)
+#14;EPSG:4326;2.114551393;50.345609791;EPSG:23032;437091.854358749;5577642.33578973;0.01
+# CTS trouve bien le même résultat que CIRCE à condition de corriger la translation
+# ED50->WGS84 qui n'est pas tout à fait la même dans le fichier prj (+towgs84=-87,-98,-121)
+# et dans circe (+towgs84=-84,-97,-117)
+14;EPSG:4326;2.114551393;50.345609791;EPSG:23032;10386.009;5599964.321;0.01
+15;EPSG:4326;2.114551393;50.345609791;EPSG:32630;863820.830208352;5589575.52381854;0.01
+16;EPSG:4326;2.114551393;50.345609791;EPSG:32631;436999.022430445;5577433.14063792;0.01
+17;EPSG:4326;2.114551393;50.345609791;EPSG:32632;10306.7238895355;5599761.98610375;0.01
+18;EPSG:4326;2.114551393;50.345609791;EPSG:3942;1636274.59149815;2131307.97042943;0.01
+19;EPSG:27582;584173.736059813;2594514.82833411;EPSG:4326;2.114551393;50.345609791;0.01
+20;EPSG:27572;584173.736059813;2594514.82833411;EPSG:4326;2.114551393;50.345609791;0.01
+21;EPSG:2154;636890.74032145;7027895.26344997;EPSG:4326;2.114551393;50.345609791;0.01
+22;EPSG:27572;282331;2273699.7;EPSG:2154;332602.961893497;6709788.26447893;0.01
+23;EPSG:2154;332602.961893497;6709788.26447893;EPSG:27572;282331;2273699.7;0.01
+24;EPSG:4326;2.114551393;50.345609791;EPSG:3857;235390.784324967;6506345.90578986;0.01
+25;EPSG:27572;282331; 2273699.7;EPSG:3857;-208496.537435372; 6005369.87702729;0.01
